Move the "draw_to_hand" link up
The "draw to hand" link has been moved above the confusingly named
"draw" link, and the latter has been renamed to "play directly from
deck".  The objective is to reduce the possibility of user error
in Chaos in the Old World games, where only the GM will ever need
to play a card off a deck; the option remains available to users
if needed for other games.